RDKCOM-2907: REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel an...
Reason for change: to match datatype compatibilty
Test Procedure: Build should be successful. mediaframework should build fine.
Risks: Low/None.

(cherry picked from commit add3a86ffaa5f75b49b8868fff021b77e3930bd0)

+++ qamsource/common/rmf_sectionfilter.h
@@ -382,7 +382,7 @@ private:
 	uint32_t getLowestPriorityFilter(uint32_t filterPriority);
 	void getTimeInMilliSecond(unsigned long long *pMillisSinceEpoch);
 	rmf_Error dispatchEvent(rmf_osal_eventqueue_handle_t queueHandle, uint32_t filterEvent, uint8_t priority, uint32_t filterID);
-	rmf_Error ReleaseFilterRequest(uint32_t uiFilterReqHandle);
+	rmf_Error ReleaseFilterRequest(uintptr_t uiFilterReqHandle);
 	rmf_Error ReleaseFromDeferredDeleteQueue();
 	bool IsDeferredDeleteQueueEmpty();
 };
